Refactor launcher injection
